Move to utils and rename to .tsx to make it work with parser.
