+++ react/components/EditorContainer/Sidebar/ConfigurationList/ContentEditor/ConditionControls/ScopeSelector/ScopeSelector.tsx
@@ -1,5 +1,5 @@
 import React, { Fragment } from 'react'
-import { defineMessages, InjectedIntlProps, injectIntl } from 'react-intl'
+import { InjectedIntlProps, injectIntl } from 'react-intl'
 import { RadioGroup } from 'vtex.styleguide'
 
 import ConditionTitle from '../ConditionTitle'
@@ -14,50 +14,6 @@ interface CustomProps {
 }
 
 type Props = CustomProps & InjectedIntlProps
-
-// Messages used in getScopeStandardOptions
-defineMessages({
-  brand: {
-    defaultMessage: 'this brand',
-    id: 'admin/pages.editor.components.condition.scope.entity.brand',
-  },
-  category: {
-    defaultMessage: 'this category',
-    id: 'admin/pages.editor.components.condition.scope.entity.category',
-  },
-  department: {
-    defaultMessage: 'this department',
-    id: 'admin/pages.editor.components.condition.scope.entity.department',
-  },
-  product: {
-    defaultMessage: 'this product',
-    id: 'admin/pages.editor.components.condition.scope.entity.product',
-  },
-  route: {
-    defaultMessage: 'this URL',
-    id: 'admin/pages.editor.components.condition.scope.entity.route',
-  },
-  search: {
-    defaultMessage: 'this search',
-    id: 'admin/pages.editor.components.condition.scope.entity.search',
-  },
-  sitewide: {
-    defaultMessage: 'the entire site',
-    id: 'admin/pages.editor.components.condition.scope.sitewide',
-  },
-  subcategory: {
-    defaultMessage: 'this subcategory',
-    id: 'admin/pages.editor.components.condition.scope.entity.subcategory',
-  },
-  template: {
-    defaultMessage: 'this template',
-    id: 'admin/pages.editor.components.condition.scope.template',
-  },
-  title: {
-    defaultMessage: 'Apply to',
-    id: 'admin/pages.editor.components.condition.scope.title',
-  },
-})
 
 const ScopeSelector: React.FunctionComponent<Props> = ({
   intl,
