--- react/components/EditorContainer/Sidebar/ConfigurationList/ContentEditor/ConditionControls/ScopeSelector/utils.ts
@@ -1,19 +1,0 @@
-import { InjectedIntl } from 'react-intl'
-
-export const getScopeStandardOptions = (
-  intl: InjectedIntl,
-  pageContext: PageContext
-) => [
-  {
-    label: intl.formatMessage({
-      id: `admin/pages.editor.components.condition.scope.entity.${pageContext.type}`,
-    }),
-    value: 'entity',
-  },
-  {
-    label: intl.formatMessage({
-      id: 'admin/pages.editor.components.condition.scope.template',
-    }),
-    value: 'template',
-  },
-]
